Таблица вынесена в отдельный компонент
